+++ userextensions/__init__.py
@@ -7,11 +7,11 @@ A user extension module for django.
 """
 
 __title__ = 'django-userextensions'
-__version__ = '0.0.5'
+__version__ = '0.0.6'
 __author__ = 'David Slusser'
 __email__ = 'dbslusser@gmail.com'
 __license__ = 'GPL-3.0'
 __copyright__ = 'Copyright 2019 David Slusser'
 
 
-default_app_config = 'userextensions.apps.userextensionsConfig'
+default_app_config = 'userextensions.apps.UserExtensionsConfig'