+++ userextensions/views/__init__.py
@@ -1,4 +1,2 @@
 from .action import *
-from .ajax import *
-from .api import *
 from .gui import *