+++ docs/source/version_history.rst
@@ -1,6 +1,20 @@
+.. _version_history:
+
+
 Version History
 ===============
 
-0.0.1
-^^^^^
 
+Pre-Release
+-----------
+
+.. csv-table::
+   :header: "Release", "Details"
+   :widths: 20, 100
+
+   "0.0.6", "added documentation via sphinx"
+   "0.0.5", "fixed base_tempalte in recents/favorites views"
+   "0.0.4", "added views and templates for user pages"
+   "0.0.3", "added start page to UserPreferences"
+   "0.0.2", "updated versions"
+   "0.0.1", "initial release"